requirements.txt
eth-brownie>=1.9.0,<2.0.0
aiohttp==3.7.4.post0
asttokens==2.0.4
async-timeout==3.0.1
attrs==21.2.0
base58==2.1.0
bitarray==1.2.2
black==21.9b0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
Cython==0.29.24
cytoolz==0.11.0
dataclassy==0.10.4
eip712==0.1.0
eth-abi==2.1.1
eth-account==0.5.5
eth-brownie==1.16.4
eth-event==1.2.3
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
execnet==1.9.0
hexbytes==0.2.2
hypothesis==6.21.6
idna==3.2
inflection==0.5.0
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
lazy-object-proxy==1.6.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==5.1.0
mypy-extensions==0.4.3
mythx-models==1.9.1
netaddr==0.8.0
packaging==21.0
parsimonious==0.8.1
pathspec==0.9.0
platformdirs==2.3.0
pluggy==1.0.0
prompt-toolkit==3.0.20
protobuf==3.18.0
psutil==5.8.0
py==1.10.0
py-solc-ast==1.2.9
py-solc-x==1.1.0
pycryptodome==3.10.1
Pygments==2.10.0
pygments-lexer-solidity==0.7.0
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
pytest-forked==1.3.0
pytest-xdist==1.34.0
python-dateutil==2.8.1
python-dotenv==0.16.0
pythx==1.6.1
PyYAML==5.4.1
regex==2021.8.28
requests==2.26.0
rlp==1.2.0
semantic-version==2.8.5
six==1.16.0
sortedcontainers==2.4.0
toml==0.10.2
tomli==1.2.1
toolz==0.11.1
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.6
varint==1.0.2
vvm==0.1.0
vyper==0.2.16
wcwidth==0.2.5
web3==5.23.1
websockets==9.1
wrapt==1.12.1
yarl==1.6.3
